#263 [chore] DetailRuleGraph로 네이밍 변경
