+++ app/src/main/java/hous/release/android/presentation/our_rules/OurRulesActivity.kt
@@ -8,6 +8,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import hous.release.android.presentation.our_rules.screen.graph.RuleNavGraph
 import hous.release.android.util.HousLogEvent.SCREEN_RULES
 import hous.release.android.util.HousLogEvent.enterScreenLogEvent
+import hous.release.designsystem.theme.HousTheme
 
 @AndroidEntryPoint
 class OurRulesActivity : ComponentActivity() {
@@ -16,7 +17,9 @@ class OurRulesActivity : ComponentActivity() {
         enterScreenLogEvent(SCREEN_RULES, javaClass.name)
         setContent {
             val navController = rememberNavController()
-            RuleNavGraph(navController)
+            HousTheme {
+                RuleNavGraph(navController)
+            }
         }
     }
 }
